requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
coreapi~=2.3.3
django-cors-headers~=4.2.0
Pillow~=10.0.0
django-filter~=23.2
cryptography~=41.0.2
requests~=2.31.0